+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented here.
 
 ## v0.8.0 - 2022-12-30
 ### Added
+- Shaders can now use up to 16 bound textures simultaneously.
 - Some refactoring and QoL improvements to ease the development of custom [IEffectRenderer](src/main/java/de/teragam/jfxshader/IEffectRenderer.java).
 - The generated texture coordinates for the shader input textures can now be accessed in the [ShaderEffectPeer](src/main/java/de/teragam/jfxshader/ShaderEffectPeer.java). This allows the shaders to compensate for the varying texture coordinates caused by JavaFX's internal texture pooling.
 - The `ImagePoolPolicy.QUANTIZED` option has been added to allow better VRAM usage while ensuring that the target texture dimensions do not fluctuate between frames.
 - Added option to invert the mask in the [BlendShapes](src/main/java/de/teragam/jfxshader/samples/blendshapes/BlendShapesEffectPeer.java) example.
 
 ### Fixed
-- Fixed an issue that did not allow the use of all 4 textures in the shaders.
+- Fixed an issue that did not allow the use of more than 2 textures in a shader.
 
 ## v0.7.0 - 2022-11-28